Fixed bugs with marker creation

+++ lib/features/game/presentation/bloc/game/game_event.dart
@@ -97,7 +97,7 @@ final class GameTapEvent extends GameEvent {
 }
 
 final class GameUpdateSpeedEvent extends GameEvent {
-  final int seconds;
+  final double seconds;
   final int tapInSeconds;
   const GameUpdateSpeedEvent({
     required this.seconds,
